refactor: names don't shadow existing
