--- weaviate/tokenize/__init__.py
@@ -1,7 +1,0 @@
-"""Module for tokenize operations."""
-
-from .async_ import _TokenizeAsync
-from .sync import _Tokenize
-from .types import TokenizeResult
-
-__all__ = ["_Tokenize", "_TokenizeAsync", "TokenizeResult"]
